+++ index.html
@@ -232,6 +232,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
   <script src="/js/functions.js"> </script>
-  <script src="/js/main.js"> </script>
+  <script src="/js/main.js" type="module"> </script>
 </body>
 </html>
